+++ tests/TaskSerializer/TaskSerializer.test.ts
@@ -3,10 +3,11 @@
  */
 
 import moment from 'moment';
-import { Task, TaskRegularExpressions } from '../../src/Task/Task';
+import { Task } from '../../src/Task/Task';
 import type { TaskDetails, TaskSerializer } from '../../src/TaskSerializer';
 import { TaskBuilder } from '../TestingTools/TaskBuilder';
 import { Priority } from '../../src/Task/Priority';
+import { TaskRegularExpressions } from '../../src/Task/TaskRegularExpressions';
 
 jest.mock('obsidian');
 window.moment = moment;
